refactor: move recorder under runtime pkg (fix split package #52)
